requirements.txt
# ── LLM / Ollama ─────────────────────────────────────────────────────────────
ollama
requests

# ── Embeddings & Vector DB ────────────────────────────────────────────────────
sentence-transformers
faiss-cpu

# ── AST / Static analysis ─────────────────────────────────────────────────────
tree-sitter
radon

# ── Profiling & system metrics ────────────────────────────────────────────────
psutil

# ── Git operations ────────────────────────────────────────────────────────────
gitpython

# ── CLI & terminal UI ─────────────────────────────────────────────────────────
click
rich

# ── Data & validation ─────────────────────────────────────────────────────────
pydantic

# ── Testing ───────────────────────────────────────────────────────────────────
pytest
pytest-benchmark

# ── Numerical ────────────────────────────────────────────────────────────────
numpy
